+++ src/app/(auth)/signin/page.tsx
@@ -1,12 +1,12 @@
 'use client';
 
 import { SubmitHandler, useForm } from 'react-hook-form';
-import { Logo } from '@/components/specific-feature/AuthPage/Logo';
-import { EmailInput } from '@/components/specific-feature/AuthPage/EmailInput';
-import { PasswordInput } from '@/components/specific-feature/AuthPage/PasswordInput';
-import { AuthFooter } from '@/components/specific-feature/AuthPage/AuthFooter';
+import { Logo } from '@/components/AuthPage/Logo';
+import { EmailInput } from '@/components/AuthPage/EmailInput';
+import { PasswordInput } from '@/components/AuthPage/PasswordInput';
+import { AuthFooter } from '@/components/AuthPage/AuthFooter';
 import { AUTH_FOOTER_MESSAGES } from '@/constants/AuthFooterMessages';
-import { MetaData } from '@/components/specific-feature/AuthPage/MetaData';
+import { MetaData } from '@/components/AuthPage/MetaData';
 import { AuthDataType } from '@/types/AuthType';
 
 export default function Signin() {
