+++ src/components/AuthPage/AuthFooter/index.tsx
