+++ src/components/AuthPage/EmailInput/index.tsx
