+++ src/components/AuthPage/Logo/index.tsx
