+++ src/components/AuthPage/NameInput/index.tsx
