+++ src/components/AuthPage/PasswordChkInput/index.tsx
