+++ src/components/AuthPage/PasswordInput/index.tsx
