+++ src/components/AuthPage/VisibilityIcon/index.tsx
